code cleanup

+++ scripts/createAdmin.js
@@ -8,11 +8,17 @@ dotenv.config();
 
 mongoose.connect(process.env.MONGO_URI)
   .then(async () => {
-    const email = 'admin@benjerrys.com';
-    const password = 'superveilig123';
+    const email = process.env.ADMIN_EMAIL;
+    const password = process.env.ADMIN_PASSWORD;
+
+    if (!email || !password) {
+      console.error('❌ ADMIN_EMAIL of ADMIN_PASSWORD ontbreekt in .env');
+      process.exit(1);
+    }
+
     const passwordHash = await bcrypt.hash(password, 10);
 
-    await Admin.deleteMany({}); // 🧹 optioneel: oude admins verwijderen
+    await Admin.deleteMany({}); 
     await Admin.create({ email, passwordHash });
 
     console.log('✅ Admin user aangemaakt!');
